Blood Donation Website/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-bootstrap5==12.0.0
django-crispy-forms==1.11.2